System.Console.ANSI.Types

Description

Types used to represent SELECT GRAPHIC RENDITION (SGR) aspects.

Documentation

data SGR

ANSI Select Graphic Rendition command

Constructors

Reset

SetConsoleIntensity !ConsoleIntensity

SetItalicized !Bool
Not widely supported: sometimes treated as swapping foreground and background

SetUnderlining !Underlining

SetBlinkSpeed !BlinkSpeed

SetVisible !Bool
Not widely supported

SetSwapForegroundBackground !Bool

SetColor !ConsoleLayer !ColorIntensity !Color

SetRGBColor !ConsoleLayer !(Colour Float)
Supported from Windows 10 Creators Update

data ConsoleLayer

ANSI colors can be set on two different layers

Constructors

Foreground

Background

data Color

ANSI colors: come in various intensities, which are controlled by ColorIntensity

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

data ColorIntensity

ANSI colors come in two intensities

Constructors

Dull

Vivid

data ConsoleIntensity

ANSI general console intensity: usually treated as setting the font style (e.g. BoldIntensity causes text to be bold)

Constructors

BoldIntensity

FaintIntensity
Not widely supported: sometimes treated as concealing text

NormalIntensity

data Underlining

ANSI text underlining

Constructors

SingleUnderline

DoubleUnderline
Not widely supported

NoUnderline

data BlinkSpeed

ANSI blink speeds: values other than NoBlink are not widely supported

Constructors

SlowBlink
Less than 150 blinks per minute

RapidBlink
More than 150 blinks per minute

NoBlink